+++ redmine/client.go
@@ -16,15 +16,19 @@ type Config struct {
 	Username       string
 	Password       string
 	SkipCertVerify bool
-	APIKey         string
 }
 
-func NewClient(config Config) *Client {
-	redmineAPI := rmapi.NewClient(config.URL, config.APIKey)
-	redmineAPI.Limit = -1
-	redmineAPI.Offset = -1
+func NewClient(config Config) (*Client, error) {
+	redmineAPI, err := rmapi.NewClientBuilder().
+		Endpoint(config.URL).
+		AuthBasicAuth(config.Username, config.Password).
+		SkipSSLVerify(config.SkipCertVerify).
+		Build()
+	if err != nil {
+		return nil, err
+	}
 
-	return &Client{config: config, redmineAPI: redmineAPI}
+	return &Client{config: config, redmineAPI: redmineAPI}, nil
 }
 
 func verifyIDtoInt(id string) (int, error) {