+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Add
+- set Basic Authentication as default authentication instead of API token (#3)
+- add support for [issue priority](docs/usage_en.md) to issue resource (#3)
+- add support for [issue category](docs/usage_en.md) to issue resource (#3) 
+- add support for [version](docs/usage_en.md) to issue resource (#3) 
 
 ## [v0.1.0] - 2021-03-05
 ### Added
-* initial provider that supports projects and issues (#1)
+- initial provider that supports projects and issues (#1)